refactor: totalElements, hasNext, favorite 타입 변경

+++ src/main/java/com/findex/team02/indexInfo/service/BasicIndexInfoService.java
@@ -40,7 +40,7 @@ public class BasicIndexInfoService implements IndexInfoService {
                 last.getId().toString(),
                 last.getId(),
                 size,
-                indexInfos.size(),
+                (long) indexInfos.size(),
                 hasNext
         );
     }
